+++ tests/regressions/hrs.ml
@@ -7,4 +7,4 @@ let () =
   Library.set_lib_root (Some "/tmp");
   Timed.(Console.verbose := 0);
   let sign = Compile.PureUpToSign.compile_file "../OK/bool.lp" in
-  Export.Hrs.to_HRS Format.std_formatter sign
+  Export.Hrs.sign Format.std_formatter sign